Export issues with Blender 3.1 (#382)
* Blender 3.1 bug with width

* Width bug with blender 3.10

* Update armature_manual.py

* Update atlas.py

* Update common.py

* Update importer.py

* Update armature.py

+++ tools/armature_manual.py
@@ -360,7 +360,7 @@ class PoseNamePopup(bpy.types.Operator):
     def invoke(self, context, event):
         context.scene.pose_to_shapekey_name = 'Pose'
         dpi_value = Common.get_user_preferences().system.dpi
-        return context.window_manager.invoke_props_dialog(self, width=dpi_value * 4)
+        return context.window_manager.invoke_props_dialog(self, width=int(dpi_value * 4))
 
     def check(self, context):
         # Important for changing options
